feat: 계산기 패키지 생성 및 기본 파일 추가
