+++ src/calculator/App.java
@@ -1,0 +1,4 @@
+package calculator;
+
+public class App {
+}
